+++ src/main/java/io/ep2p/kademlia/netty/serialization/KademliaMessageDeserializer.java
@@ -11,26 +11,27 @@ import lombok.SneakyThrows;
 
 import java.io.Serializable;
 import java.lang.reflect.Type;
+import java.math.BigInteger;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 
-public class KademliaMessageDeserializer<K extends Serializable, V extends Serializable> implements JsonDeserializer<KademliaMessage<Long, NettyConnectionInfo, Serializable>> {
+public class KademliaMessageDeserializer<K extends Serializable, V extends Serializable> implements JsonDeserializer<KademliaMessage<BigInteger, NettyConnectionInfo, Serializable>> {
     private final Map<String, Type> typeRegistry = new ConcurrentHashMap<>();
     private final Map<String, Class<?>> messageClassRegistry = new ConcurrentHashMap<>();
 
     public KademliaMessageDeserializer() {
-        this.registerDataType(MessageType.DHT_LOOKUP, new TypeToken<DHTLookupKademliaMessage.DHTLookup<Long, NettyConnectionInfo, K>>(){}.getType());
+        this.registerDataType(MessageType.DHT_LOOKUP, new TypeToken<DHTLookupKademliaMessage.DHTLookup<BigInteger, NettyConnectionInfo, K>>(){}.getType());
         this.registerMessageClass(MessageType.DHT_LOOKUP, DHTLookupKademliaMessage.class);
         this.registerDataType(MessageType.DHT_LOOKUP_RESULT, new TypeToken<DHTLookupResultKademliaMessage.DHTLookupResult<K, V>>(){}.getType());
         this.registerMessageClass(MessageType.DHT_LOOKUP_RESULT, DHTLookupResultKademliaMessage.class);
-        this.registerDataType(MessageType.DHT_STORE, new TypeToken<DHTStoreKademliaMessage.DHTData<Long, NettyConnectionInfo, K, V>>(){}.getType());
+        this.registerDataType(MessageType.DHT_STORE, new TypeToken<DHTStoreKademliaMessage.DHTData<BigInteger, NettyConnectionInfo, K, V>>(){}.getType());
         this.registerMessageClass(MessageType.DHT_STORE, DHTStoreKademliaMessage.class);
         this.registerDataType(MessageType.DHT_STORE_RESULT, new TypeToken<DHTStoreResultKademliaMessage.DHTStoreResult<K>>(){}.getType());
         this.registerMessageClass(MessageType.DHT_STORE_RESULT, DHTStoreResultKademliaMessage.class);
-        this.registerDataType(MessageType.FIND_NODE_REQ, new TypeToken<Long>(){}.getType());
+        this.registerDataType(MessageType.FIND_NODE_REQ, new TypeToken<BigInteger>(){}.getType());
         this.registerMessageClass(MessageType.FIND_NODE_REQ, FindNodeRequestMessage.class);
-        this.registerDataType(MessageType.FIND_NODE_RES, new TypeToken<FindNodeAnswer<Long, NettyConnectionInfo>>(){}.getType());
+        this.registerDataType(MessageType.FIND_NODE_RES, new TypeToken<FindNodeAnswer<BigInteger, NettyConnectionInfo>>(){}.getType());
         this.registerMessageClass(MessageType.FIND_NODE_RES, FindNodeResponseMessage.class);
         this.registerDataType(MessageType.PING, new TypeToken<String>(){}.getType());
         this.registerMessageClass(MessageType.PING, PingKademliaMessage.class);
@@ -44,17 +45,17 @@ public class KademliaMessageDeserializer<K extends Serializable, V extends Seria
 
     @SneakyThrows
     @Override
-    public KademliaMessage<Long, NettyConnectionInfo, Serializable> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
+    public KademliaMessage<BigInteger, NettyConnectionInfo, Serializable> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
 
         JsonObject jsonObject = jsonElement.getAsJsonObject();
         String messageType = jsonObject.getAsJsonPrimitive("type").getAsString();
-        Node<Long, NettyConnectionInfo> node = jsonDeserializationContext.deserialize(
+        Node<BigInteger, NettyConnectionInfo> node = jsonDeserializationContext.deserialize(
                 jsonObject.getAsJsonObject("node"),
                 Node.class
         );
         Class<?> aClass = this.messageClassRegistry.get(messageType);
         @SuppressWarnings("unchecked")
-        KademliaMessage<Long, NettyConnectionInfo, Serializable> o = (KademliaMessage<Long, NettyConnectionInfo, Serializable>) aClass.getConstructor().newInstance();
+        KademliaMessage<BigInteger, NettyConnectionInfo, Serializable> o = (KademliaMessage<BigInteger, NettyConnectionInfo, Serializable>) aClass.getConstructor().newInstance();
         o.setData(getData(messageType, jsonObject, jsonDeserializationContext));
         o.setType(messageType);
         o.setNode(node);
