+++ src/main/java/io/ep2p/kademlia/netty/serialization/NodeInstanceCreator.java
@@ -6,11 +6,12 @@ import io.ep2p.kademlia.netty.common.NettyExternalNode;
 import io.ep2p.kademlia.node.Node;
 
 import java.lang.reflect.Type;
+import java.math.BigInteger;
 
 
-public class NodeInstanceCreator implements InstanceCreator<Node<Long, NettyConnectionInfo>> {
+public class NodeInstanceCreator implements InstanceCreator<Node<BigInteger, NettyConnectionInfo>> {
     @Override
-    public Node<Long, NettyConnectionInfo> createInstance(Type type) {
+    public Node<BigInteger, NettyConnectionInfo> createInstance(Type type) {
         return new NettyExternalNode();
     }
 }
