+++ src/test/java/io/ep2p/kademlia/netty/FilterChainTest.java
@@ -20,6 +20,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.math.BigInteger;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
@@ -40,20 +41,20 @@ public class FilterChainTest {
         NodeSettings.Default.BUCKET_SIZE = 100;
         NodeSettings.Default.PING_SCHEDULE_TIME_VALUE = 5;
 
-        KeyHashGenerator<Long, String> keyHashGenerator = key -> {
+        KeyHashGenerator<BigInteger, String> keyHashGenerator = key -> {
             try {
-                return new BoundedHashUtil(NodeSettings.Default.IDENTIFIER_SIZE).hash(key.hashCode(), Long.class);
+                return new BoundedHashUtil(NodeSettings.Default.IDENTIFIER_SIZE).hash(key.hashCode(), BigInteger.class);
             } catch (UnsupportedBoundingException e) {
                 e.printStackTrace();
             }
-            return Long.valueOf(key.hashCode());
+            return BigInteger.valueOf(key.hashCode());
         };
 
         okHttpMessageSender1 = new OkHttpMessageSender<>();
 
         // node 1
         node1 = new NettyKademliaDHTNodeBuilder<>(
-                1L,
+                BigInteger.valueOf(1L),
                 new NettyConnectionInfo("127.0.0.1", NodeHelper.findRandomPort()),
                 new SampleRepository(),
                 keyHashGenerator
@@ -80,7 +81,7 @@ public class FilterChainTest {
 
 
         NettyKademliaDHTNode<String, String> node2 = new NettyKademliaDHTNodeBuilder<>(
-                2L,
+                BigInteger.valueOf(2L),
                 new NettyConnectionInfo("127.0.0.1", NodeHelper.findRandomPort()),
                 new SampleRepository(),
                 node1.getKeyHashGenerator()
